requirements.txt
Flask==0.12.1
Flask-Cors==3.0.2
python-docx
requests
ipfsapi
pdfkit